pyproject.toml
[project]
name = "workinkorea-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.2",
    "fastapi-mail>=1.5.0",
    "httpx>=0.28.1",
    "minio>=7.2.18",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.20",
    "redis>=7.0.0",
    "sqlalchemy[asyncio]>=2.0.43",
    "uvicorn>=0.35.0",
]
